feat: condense useful implementation files
Deduplicate the various implementation files and move them into a common
directory. Use a naming scheme which allows all chosen implementations
to live next to each other.

Later on we will deduplicate function implementations, but this is a
good intermediate state which should significantly ease tracing the
meaningful modifications of the refactoring.