feat: implement runtime implementation dispatching
Add a central registry called `dplx_blake2_impl_vtables` which collects
a vtable for each implementation. The vtable consists of all public APIs
of an implementation. Additionally define an internal function pointer
for each public API which represents the currently selected
implementation for said API. Provide functions for manually inspecting
available implementations and selecting one. Said internal function
pointers are initialized with a dummy dispatcher function which lazily
tries to select the "best" fitting implementation.

The system should be easily extendable to utilize IFUNCS or
alternatively avoid lazy initialization by using a global constructor.

The dispatcher implementation is replaced by a stub if the library is
compiled with a single implementation. In this case the implementation
is directly exposed as the public API.